refactor(api): move route in src/organizational/entities
for api/admin/tags
