+++ api/src/organizational-entities/application/routes.js
@@ -1,6 +1,7 @@
 import * as certificationCenterAdminRoutes from './certification-center/certification-center.admin.route.js';
 import * as organizationAdminRoutes from './organization/organization.admin.route.js';
+import * as tagAdminRoutes from './tag/tag.admin.route.js';
 
-const organizationalEntitiesRoutes = [organizationAdminRoutes, certificationCenterAdminRoutes];
+const organizationalEntitiesRoutes = [certificationCenterAdminRoutes, organizationAdminRoutes, tagAdminRoutes];
 
 export { organizationalEntitiesRoutes };
