+++ api/tests/acceptance/application/tags/tag-api_test.js
@@ -10,64 +10,6 @@ import {
 } from '../../../test-helper.js';
 
 describe('Acceptance | Route | tag-router', function () {
-  describe('POST /api/admin/tags', function () {
-    it('should return the created tag with 201 HTTP status code', async function () {
-      // given
-      const tagName = 'SUPER TAG';
-      const server = await createServer();
-      await databaseBuilder.commit();
-
-      const userId = (await insertUserWithRoleSuperAdmin()).id;
-
-      // when
-      const response = await server.inject({
-        method: 'POST',
-        url: '/api/admin/tags',
-        payload: {
-          data: {
-            type: 'tags',
-            attributes: {
-              name: tagName,
-            },
-          },
-        },
-        headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
-      });
-
-      // then
-      expect(response.statusCode).to.equal(201);
-      expect(response.result.data.type).to.deep.equal('tags');
-      expect(response.result.data.attributes.name).to.deep.equal(tagName);
-    });
-
-    it('should return 403 HTTP status code when the user authenticated is not SuperAdmin', async function () {
-      // given
-      const server = await createServer();
-      const userId = databaseBuilder.factory.buildUser().id;
-      await databaseBuilder.commit();
-
-      const tagName = 'un super tag';
-
-      // when
-      const response = await server.inject({
-        method: 'POST',
-        url: '/api/admin/tags',
-        payload: {
-          data: {
-            type: 'tags',
-            attributes: {
-              name: tagName,
-            },
-          },
-        },
-        headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
-      });
-
-      // then
-      expect(response.statusCode).to.equal(403);
-    });
-  });
-
   describe('GET /api/admin/tags', function () {
     it('should return a list of tags with 200 HTTP status code', async function () {
       // given
