+++ api/tests/unit/application/tags/index_test.js
@@ -26,35 +26,4 @@ describe('Unit | Application | Router | tag-router', function () {
       sinon.assert.calledOnce(tagController.findAllTags);
     });
   });
-
-  describe('POST /api/admin/tags', function () {
-    it('returns forbidden access if admin member has CERTIF or SUPPORT or METIER role', async function () {
-      // given
-      sinon.stub(tagController, 'create').resolves('ok');
-
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response().code(403).takeover());
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleCertif').callsFake((request, h) => h.response(true));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport').callsFake((request, h) => h.response(true));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier').callsFake((request, h) => h.response(true));
-
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('POST', '/api/admin/tags', {
-        data: {
-          type: 'tags',
-          attributes: {
-            name: 'Super Tag',
-          },
-        },
-      });
-
-      // then
-      expect(response.statusCode).to.equal(403);
-      sinon.assert.notCalled(tagController.create);
-    });
-  });
 });
